+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/kubernetes-crd/compare/0.5.0...HEAD
 
+### Added
+
+- Support for setting the experiment as a YAML file [#71][71]
+
+[71]: https://github.com/chaostoolkit-incubator/kubernetes-crd/issues/71
+
 ## [0.5.0][] - 2021-06-08
 
 [0.5.0]: https://github.com/chaostoolkit-incubator/kubernetes-crd/compare/0.3.4...0.5.0